Refactor sponsor access
