+++ src/firebase/sponsors/reload-all.js
@@ -1,0 +1,5 @@
+const deleteAllSponsors = require('./delete-all')
+const setAllSponsors = require('./set-all')
+
+module.exports = firebaseAdmin =>
+  deleteAllSponsors(firebaseAdmin).then(() => setAllSponsors(firebaseAdmin))
