+++ src/firebase/sponsors/set-all.js
@@ -1,0 +1,9 @@
+const sponsors = require('../../../data/sponsors.json')
+
+module.exports = firebaseAdmin => {
+  const sponsorsRef = firebaseAdmin.firestore().collection('sponsors')
+  const promises = Object.keys(sponsors).map(key =>
+    sponsorsRef.doc(key).set(sponsors[key])
+  )
+  return Promise.all(promises)
+}
